Misc stylistic changes
