--- Code/EditorPlugins/Assets/EditorPluginAssets/CustomDataAsset/CustomDataAssetProperties.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include <Core/Utils/CustomData.h>
-
-
-class ezCustomDataAssetProperties : public ezReflectedClass
-{
-  EZ_ADD_DYNAMIC_REFLECTION(ezCustomDataAssetProperties, ezReflectedClass);
-public:
-  ezCustomData* m_pType = nullptr;
-};
-
