+++ Code/Engine/Core/Utils/CustomData.h
@@ -18,9 +18,6 @@ public:
   virtual void Save(ezStreamWriter& inout_stream) const;
 };
 
-
-// \todo A macro creating a derived type from ezCustomDataResource
-
 template<typename T>
 class ezCustomDataResource : public ezResource
 {
@@ -43,77 +40,20 @@ private:
   T m_Data;
 };
 
+#define EZ_DECLARE_CUSTOM_DATA(SELF) \
+  class SELF##Resource : public ezCustomDataResource<SELF> \
+  { \
+  EZ_ADD_DYNAMIC_REFLECTION(SELF##Resource, ezCustomDataResource<SELF>); \
+  EZ_RESOURCE_DECLARE_COMMON_CODE(SELF##Resource); \
+  }; \
+  \
+  using SELF##ResourceHandle = ezTypedResourceHandle<SELF##Resource>; \
 
-// \todo Move everything below to a separate _inl.h file
+#define EZ_DEFINE_CUSTOM_DATA(SELF) \
+  EZ_BEGIN_DYNAMIC_REFLECTED_TYPE(SELF##Resource, 1, ezRTTIDefaultAllocator<SELF##Resource>) \
+  EZ_END_DYNAMIC_REFLECTED_TYPE; \
+  \
+  EZ_RESOURCE_IMPLEMENT_COMMON_CODE(SELF##Resource);\
 
-template<typename T>
-ezCustomDataResource<T>::ezCustomDataResource() : ezResource(DoUpdate::OnAnyThread, 1)
-{
-}
 
-template<typename T>
-ezCustomDataResource<T>::~ezCustomDataResource() = default;
-
-template<typename T>
-ezResourceLoadDesc ezCustomDataResource<T>::UnloadData(Unload WhatToUnload)
-{
-  ezResourceLoadDesc res;
-  res.m_uiQualityLevelsDiscardable = 0;
-  res.m_uiQualityLevelsLoadable = 0;
-  res.m_State = ezResourceState::Unloaded;
-  return res;
-}
-
-template<typename T>
-ezResourceLoadDesc ezCustomDataResource<T>::UpdateContent(ezStreamReader* Stream)
-{
-  ezResourceLoadDesc res;
-  res.m_uiQualityLevelsDiscardable = 0;
-  res.m_uiQualityLevelsLoadable = 0;
-
-  if (Stream == nullptr)
-  {
-    res.m_State = ezResourceState::LoadedResourceMissing;
-    return res;
-  }
-
-  // skip the absolute file path data that the standard file reader writes into the stream
-  {
-    ezStringBuilder sAbsFilePath;
-    (*Stream) >> sAbsFilePath;
-  }
-
-  ezAssetFileHeader AssetHash;
-  AssetHash.Read(*Stream).IgnoreResult();
-
-  {
-    T dummy;
-    dummy.Load(*Stream);
-
-    CreateResource(std::move(dummy));
-  }
-
-  res.m_State = ezResourceState::Loaded;
-  return res;
-}
-
-template<typename T>
-void ezCustomDataResource<T>::UpdateMemoryUsage(MemoryUsage& out_NewMemoryUsage)
-{
-  out_NewMemoryUsage.m_uiMemoryCPU = sizeof(ezCustomDataResource<T>);
-  out_NewMemoryUsage.m_uiMemoryGPU = 0;
-}
-
-template<typename T>
-EZ_RESOURCE_IMPLEMENT_CREATEABLE(ezCustomDataResource<T>, T)
-{
-  m_Data = descriptor;
-
-  ezResourceLoadDesc res;
-  res.m_uiQualityLevelsDiscardable = 0;
-  res.m_uiQualityLevelsLoadable = 0;
-  res.m_State = ezResourceState::Loaded;
-
-  return res;
-}
-
+#include <Core/Utils/Implementation/IntervalScheduler_inl.h>
